Updating blacklist & help!

+++ main.js
@@ -243,6 +243,7 @@ client.on('message', message => {
 });
 
 client.on('message', message => {
+    if (message.guild === null) return;
     if (message.type === "PINS_ADD" && message.author.bot) return message.delete();
 
     if (message.content) {
@@ -261,15 +262,12 @@ client.on('message', message => {
     if (message.channel.type !== "dm") return;
 
     if (message.content.startsWith(`+blacklist`) || message.content.startsWith(`+blocklist`)) {
-        const blacklistEmbed = new Discord.MessageEmbed()
-            .setAuthor(`Blacklisted Words`, message.author.displayAvatarURL({
-                dynamic: true
-            }))
-            .setDescription(`**📜 ${message.author.username}**, the list of blacklisted words can be found by [clicking on here](https://i.imgur.com/0r0kU7O.jpg), or viewing the image below.`)
-            .setImage('https://i.imgur.com/0r0kU7O.jpg')
+        message.author.send(`**[📜] ${message.author.username}**, the list of blacklisted words in **I Talk Server** can be found below.`, { files: ['https://i.imgur.com/KY50hx2.png'] }).then((messageSent) => {
+            message.author.send({ files: ['https://i.imgur.com/98mgvwY.png']});
 
-        message.author.send({
-            embed: blacklistEmbed
+            setTimeout(() => {
+                messageSent.pin();
+            }, 3000);
         });
     }
 });
